feat: refactor and docker images infrence vs train
