+++ .env.dist
@@ -5,5 +5,3 @@ NEO4J_PASSWORD = "YOUR_NEO4J_PASSWORD"
 
 JWT_SECRET_KEY = "YOUR_JWT_SECRET_KEY"
 JWT_ALGORITHM  = "JWT_ALGORITHM"
-
-INFERENCE_URL = "YOUR_INFERENCE_URL"
