+++ .github/workflows/docker-inference.yml
@@ -53,7 +53,7 @@ jobs:
         with:
           context: .
           platforms: linux/amd64
-          file: .docker/Dockerfile
+          file: tools/images/Dockerfile-inference
           push: false
           tags: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}:${{ steps.release-version.outputs.version_build }}
           build-args: PLATFORM=cu124
@@ -117,7 +117,7 @@ jobs:
         with:
           context: .
           platforms: linux/amd64
-          file: .docker/Dockerfile
+          file: tools/images/Dockerfile-inference
           push: true
           tags: ${{ steps.meta.outputs.tags }}
           labels: ${{ steps.meta.outputs.labels }}
